requirements.txt
# Core
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0

# YouTube & Google
google-api-python-client>=2.100.0
pytrends>=4.9.0

# Database & Caching
sqlalchemy>=2.0.0

# Notion Integration
notion-client>=2.0.0

# CLI
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0

# Export
openpyxl>=3.1.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
